+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2](https://github.com/EnforceAuth/zift/compare/v0.1.1...v0.1.2) - 2026-04-28
+
+### Added
+
+- *(rules/java)* heuristic detection of custom authz service calls ([#13](https://github.com/EnforceAuth/zift/pull/13))
+- detect namespace, require, and destructured policy imports ([#9](https://github.com/EnforceAuth/zift/pull/9))
+
+### Fixed
+
+- *(scanner/imports)* use code-side alias for named ES imports ([#14](https://github.com/EnforceAuth/zift/pull/14))
+- *(rules/java)* match non-literal args in role/permission checks ([#12](https://github.com/EnforceAuth/zift/pull/12))
+- *(rules/java)* repair invalid query in security-interface-impl ([#11](https://github.com/EnforceAuth/zift/pull/11))
+
 ## [0.1.1](https://github.com/EnforceAuth/zift/compare/v0.1.0...v0.1.1) - 2026-04-23
 
 ### Added